day 6 part 2

+++ src/main/java/com/dan/advent/day06/Race.java
@@ -3,22 +3,23 @@ package com.dan.advent.day06;
 import lombok.RequiredArgsConstructor;
 
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 @RequiredArgsConstructor
 public class Race {
 
-    private final int time;
-    private final int record;
+    private final long time;
+    private final long record;
 
     int winningWays() {
-        return (int) IntStream.rangeClosed(1, time)
+        return (int) LongStream.rangeClosed(1, time)
                 .map(this::distance)
                 .filter(d -> d > record)
                 .count();
     }
 
-    private int distance(int hold) {
-        int moveDuration = time - hold;
+    private long distance(long hold) {
+        long moveDuration = time - hold;
         return hold * moveDuration;
     }
 }
